python-common: Move Drive Group tests
