+++ src/pybind/mgr/orchestrator_cli/requirements.txt
@@ -1,1 +1,2 @@
 tox==2.9.1
+-e ../../../python-common
