+++ src/python-common/tox.ini
@@ -1,0 +1,18 @@
+[tox]
+envlist = py27, py35, py36, flake8
+skip_missing_interpreters = true
+
+[testenv]
+deps=
+  pytest
+commands=py.test -v {posargs:ceph/tests}
+
+[testenv:flake8]
+deps=flake8
+commands=flake8 {posargs:ceph}
+
+[tool:pytest]
+norecursedirs = .* _* virtualenv
+
+[flake8]
+select=F,E9
